+++ editor/editor-core/src/__tests__/visual-regression/common/emoji.ts
@@ -37,7 +37,8 @@ describe('Emoji', () => {
       await page.click(emojiSelectors.standard);
     });
 
-    it('displays custom emoji as selected', async () => {
+    // FIXME: This test was automatically skipped due to failure on 25/07/2022: https://product-fabric.atlassian.net/browse/ED-15304
+    it.skip('displays custom emoji as selected', async () => {
       await page.click(emojiSelectors.custom);
     });
   });