+++ linking-platform/link-picker/src/ui/index.tsx
@@ -1,12 +1,14 @@
 import React, { memo } from 'react';
 import { AnalyticsErrorBoundary } from '@atlaskit/analytics-next';
+import { LazySuspense, lazyForPaint } from 'react-loosely-lazy';
 
 import {
   name as packageName,
   version as packageVersion,
 } from '../version.json';
 import { ErrorBoundaryFallback } from './error-boundary-fallback';
-import { LinkPickerWithIntl, LinkPickerProps } from './link-picker';
+import { LinkPickerProps } from './link-picker';
+import { LoaderFallback } from './loader-fallback';
 
 const COMPONENT_NAME = 'LinkPicker';
 const ANALYTICS_CHANNEL = 'media';
@@ -17,12 +19,22 @@ const PACKAGE_DATA = {
   componentName: COMPONENT_NAME,
 };
 
+const LazyLinkPicker = lazyForPaint(
+  () =>
+    import(
+      /* webpackChunkName: "@atlaskit-internal_link-picker" */
+      './link-picker'
+    ),
+);
+
 export const ComposedLinkPicker = memo((props: LinkPickerProps) => (
   <AnalyticsErrorBoundary
     channel={ANALYTICS_CHANNEL}
     data={PACKAGE_DATA}
     ErrorComponent={ErrorBoundaryFallback}
   >
-    <LinkPickerWithIntl {...props} />
+    <LazySuspense fallback={LoaderFallback}>
+      <LazyLinkPicker {...props} />
+    </LazySuspense>
   </AnalyticsErrorBoundary>
 ));