+++ linking-platform/link-picker/src/ui/loader-fallback/index.tsx
@@ -1,0 +1,14 @@
+/** @jsx jsx */
+import { jsx } from '@emotion/react';
+import Spinner from '@atlaskit/spinner';
+
+import { rootContainerStylesForLoader } from '../link-picker/styled';
+
+export const LoaderFallback = (
+  <div
+    css={rootContainerStylesForLoader}
+    data-testid="link-picker-root-loader-boundary-ui"
+  >
+    <Spinner testId="link-picker.component-loading-indicator" size="medium" />
+  </div>
+);