+++ media/smart-card/src/view/FlexibleCard/components/common/loading-skeleton/types.ts
@@ -5,4 +5,12 @@ export type LoadingSkeletonProps = {
    * code, serving as a hook for automated tests
    */
   testId?: string;
+  /**
+   * Width of the loading skeleton, defaults to width of the parent element
+   */
+  width?: number;
+  /**
+   * Height of loading skeleton, defaults to height of the parent element
+   */
+  height?: number;
 };